Adding Token to the header (id+basic+phone)

+++ src/main/java/com/menainnovations/weshare/controller/FollowerController.java
@@ -3,6 +3,7 @@ package com.menainnovations.weshare.controller;
 import com.menainnovations.weshare.model.Follower;
 import com.menainnovations.weshare.model.User;
 import com.menainnovations.weshare.services.FollowerServiceImpl;
+import com.menainnovations.weshare.validator.UserTokenValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,15 +15,29 @@ public class FollowerController {
 
     @Autowired
     FollowerServiceImpl followerService;
-
+    @Autowired
+    UserTokenValidator userTokenValidator;
     @RequestMapping(value = "/user/{userId}/followers" , method = RequestMethod.GET)
-    public String getFollower(@PathVariable long userId){
-        return followerService.getAllFollowers(userId);
+    public String getFollower(@RequestHeader("Access-Token") String token ,@PathVariable long userId){
+        if(token.trim()==""){
+            return "{\"status\": 400}";
+        }else if (userTokenValidator.validateToken(userId,token)){
+            return followerService.getAllFollowers(userId);
+        }else {
+            return "{\"status\": 400}";
+        }
     }
 
     @RequestMapping(value = "/user/{followerId}/following" , method = RequestMethod.GET)
-    public String getFollowing(@PathVariable long followerId){
-        return followerService.getAllFollowing(followerId);
+    public String getFollowing(@RequestHeader("Access-Token") String token ,@PathVariable long followerId){
+        if(token.trim()==""){
+            return "{\"status\": 400}";
+        }else if (userTokenValidator.validateToken(followerId,token)){
+            return followerService.getAllFollowing(followerId);
+        }else {
+            return "{\"status\": 400}";
+        }
+
     }
 
     /*
@@ -30,25 +45,49 @@ public class FollowerController {
      */
 
     @RequestMapping(value = "/user/{followerId}/follow/{userId}" , method = RequestMethod.GET)
-    public boolean isFollowing(@PathVariable long userId ,@PathVariable long followerId){
-        Follower follower= followerService.getTheRelationBetweenTwoUsers(followerId,userId);
-        if (follower==null){
-            return false;
+    public String isFollowing(@RequestHeader("Access-Token") String token ,@PathVariable long userId ,@PathVariable long followerId){
+        if(token.trim()==""){
+            return "{\"status\": 400}";
+        }else if (userTokenValidator.validateToken(followerId,token)){
+            Follower follower= followerService.getTheRelationBetweenTwoUsers(followerId,userId);
+            if (follower==null){
+                return "{\"status\": 0}";
+            }else {
+                return "{\"status\": 1}";
+            }
         }else {
-            return true;
+            return "{\"status\": 400}";
         }
+
+
+
     }
 
     @RequestMapping(value = "/user/{followerId}/follow/{userId}" , method = RequestMethod.POST)
-    public void addFollow(@PathVariable long followerId ,@PathVariable long userId){
-        Follower follower=new Follower();
-        follower.setFollowerId(followerId);
-        follower.setUserId(userId);
-        followerService.addFollower(follower);
+    public String addFollow(@RequestHeader("Access-Token") String token ,@PathVariable long followerId ,@PathVariable long userId){
+        if(token.trim()==""){
+            return "{\"status\": 400}";
+        }else if (userTokenValidator.validateToken(followerId,token)){
+            Follower follower=new Follower();
+            follower.setFollowerId(followerId);
+            follower.setUserId(userId);
+            followerService.addFollower(follower);
+        }else {
+            return "{\"status\": 400}";
+        }
+        return "{\"status\": 1}";
     }
 
     @RequestMapping(value = "user/{followerId}/unfollow/{userId}" , method = RequestMethod.DELETE)
-    public void deleteFollow(@PathVariable long followerId,@PathVariable long userId){
-        followerService.deleteFollower(followerId,userId);
+    public String deleteFollow(@RequestHeader("Access-Token") String token ,@PathVariable long followerId,@PathVariable long userId){
+        if(token.trim()==""){
+            return "{\"status\": 400}";
+        }else if (userTokenValidator.validateToken(followerId,token)){
+             followerService.deleteFollower(followerId,userId);
+             return "{\"status\": 1}";
+        }else {
+            return "{\"status\": 400}";
+        }
+
     }
 }
